refactor: ZLdap not abstract
- there is no point to make an abstraction if then the code casts it. ZLdapContext is strictly tied to UBID and LDAP. This removes also the need to use singletons
