+++ store/src/main/java/com/zimbra/cs/ldap/ZSearchControls.java
@@ -5,31 +5,52 @@
 
 package com.zimbra.cs.ldap;
 
-import com.zimbra.cs.ldap.unboundid.UBIDSearchControls;
+import java.util.Arrays;
+import java.util.List;
 
-public abstract class ZSearchControls {
+public class ZSearchControls {
 
     public static final int SIZE_UNLIMITED  = 0;
     public static final int TIME_UNLIMITED  = 0;
     public static final String[] RETURN_ALL_ATTRS = null;
-    
-    /*
-     * Note: In additional to the params that can be specified, all ZSearchControls 
-     * have the following properties:
-     * - no time limit
-     * - do not dereference links during search
-     */
-    
+
+    private final ZSearchScope searchScope;
+    private final int sizeLimit;
+    private final List<String> returnAttrs;
+
+    private ZSearchControls(ZSearchScope searchScope, int sizeLimit, String[] returnAttrs) {
+        this.searchScope = searchScope;
+        this.sizeLimit = sizeLimit;
+        this.returnAttrs = returnAttrs != null ? Arrays.asList(returnAttrs) : null;
+    }
+
     public static ZSearchControls SEARCH_CTLS_SUBTREE() {
         return createSearchControls(ZSearchScope.SEARCH_SCOPE_SUBTREE,
                 SIZE_UNLIMITED, RETURN_ALL_ATTRS);
     }
-    
+
     public static ZSearchControls createSearchControls(ZSearchScope searchScope,
             int sizeLimit, String[] returnAttrs) {
-        return new UBIDSearchControls(searchScope, sizeLimit, returnAttrs);
+        return new ZSearchControls(searchScope, sizeLimit, returnAttrs);
     }
 
+    public ZSearchScope getSearchScope() {
+        return searchScope;
+    }
 
+    public int getSizeLimit() {
+        return sizeLimit;
+    }
+
+    public int getTimeLimit() {
+        return TIME_UNLIMITED;
+    }
+
+    public boolean getTypesOnly() {
+        return false;
+    }
+
+    public List<String> getReturnAttrs() {
+        return returnAttrs;
+    }
 }
-
