+++ store/src/main/java/com/zimbra/cs/ldap/unboundid/UBIDLdapSchema.java
@@ -21,7 +21,7 @@ public class UBIDLdapSchema extends ZLdapSchema {
 
     private Schema schema;
 
-    UBIDLdapSchema(Schema schema) {
+    public UBIDLdapSchema(Schema schema) {
         this.schema = schema;
     }
